Performs further refactoring. Splits code up into more files. Combining all related components for a given event type into its respective file.
Adds a great deal of code comments, covering explanations
of what various types, methods and constants mean.

+++ ioctl.go
@@ -62,22 +62,21 @@ func init() {
 	sizeof_keymap_entry := int(unsafe.Sizeof(ke))
 	sizeof_effect := int(unsafe.Sizeof(ffe))
 
-	_EVIOCGVERSION = uintptr(_IOR('E', 0x01, sizeof_int))
-	_EVIOCGID = uintptr(_IOR('E', 0x02, sizeof_id))
-	_EVIOCGREP = uintptr(_IOR('E', 0x03, sizeof_int2))
-	_EVIOCSREP = uintptr(_IOW('E', 0x03, sizeof_int2))
+	_EVIOCGVERSION = _IOR('E', 0x01, sizeof_int)
+	_EVIOCGID = _IOR('E', 0x02, sizeof_id)
+	_EVIOCGREP = _IOR('E', 0x03, sizeof_int2)
+	_EVIOCSREP = _IOW('E', 0x03, sizeof_int2)
 
-	_EVIOCGKEYCODE = uintptr(_IOR('E', 0x04, sizeof_int2))
-	_EVIOCGKEYCODE_V2 = uintptr(_IOR('E', 0x04, sizeof_keymap_entry))
-	_EVIOCSKEYCODE = uintptr(_IOW('E', 0x04, sizeof_int2))
-	_EVIOCSKEYCODE_V2 = uintptr(_IOW('E', 0x04, sizeof_keymap_entry))
+	_EVIOCGKEYCODE = _IOR('E', 0x04, sizeof_int2)
+	_EVIOCGKEYCODE_V2 = _IOR('E', 0x04, sizeof_keymap_entry)
+	_EVIOCSKEYCODE = _IOW('E', 0x04, sizeof_int2)
+	_EVIOCSKEYCODE_V2 = _IOW('E', 0x04, sizeof_keymap_entry)
 
-	_EVIOCSFF = uintptr(_IOC(_IOC_WRITE, 'E', 0x80, sizeof_effect))
-	_EVIOCRMFF = uintptr(_IOW('E', 0x81, sizeof_int))
-	_EVIOCGEFFECTS = uintptr(_IOR('E', 0x84, sizeof_int))
-	_EVIOCGRAB = uintptr(_IOW('E', 0x90, sizeof_int))
-	_EVIOCSCLOCKID = uintptr(_IOW('E', 0xa0, sizeof_int))
-
+	_EVIOCSFF = _IOC(_IOC_WRITE, 'E', 0x80, sizeof_effect)
+	_EVIOCRMFF = _IOW('E', 0x81, sizeof_int)
+	_EVIOCGEFFECTS = _IOR('E', 0x84, sizeof_int)
+	_EVIOCGRAB = _IOW('E', 0x90, sizeof_int)
+	_EVIOCSCLOCKID = _IOW('E', 0xa0, sizeof_int)
 }
 
 func _EVIOCGNAME(len int) uintptr {
